bugfix: missing db column on webhook (#124)
`pullrequestinfo` table in the database contains a `merged` attribute
that takes a boolean value as to whether the PR is merged or not. I am
pretty sure it was there since I coded it myself but somehow it was
missing during the insertion. We were getting an
`Npgsql.PostgresException` that said `INSERT has more target columns
than expressions`.