fix: reply fetching suffering from race condition (#216)
If replies are not saved to DB at this point, an immediate attempt to
refetch comments will not include the latest reply. This is because the
DB insertion was taking place only on the webhook side, which is why a
race condition emerged.

It could be better to scrap the DB side altogether and embed comment
metadata into the comment body itself, however.